# docs/swagger.yaml
basePath: /
consumes:
- application/x-www-form-urlencoded
definitions:
...
  Error:
    properties:
      code:
        type: string
        x-go-name: Code
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: organizr/server/types
...
  TaskColumn:
    properties:
      board_id:
        format: int64
        type: integer
        x-go-name: BoardId
      gid:
        type: string
        x-go-name: Gid
      id:
        format: int64
        type: integer
        x-go-name: Id
      title:
        type: string
        x-go-name: Title
    type: object
    x-go-package: organizr/server/types
host: localhost:1323
info:
  description: API Documentation.
  title: Organizr
  version: 1.0.0
paths:
...
  /api/column/{Board_GID}:
    get:
      description: Retrieves all columns by parent board UUID
      operationId: column-retrieve-all
      parameters:
      - description: UUID of parent board
        in: path
        name: Board_GID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/multi-column-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - column
    post:
      description: Creates a column in the board specified by UUID
      operationId: column-create
      responses:
        "200":
          $ref: '#/responses/single-column-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - column
  /api/column/{Board_GID}/{Column_GID}:
    delete:
      description: Deletes column by parent board and column UUIDs
      operationId: column-delete
      parameters:
      - description: UUID of parent board
        in: path
        name: Board_GID
        required: true
        type: string
      - description: UUID of column
        in: path
        name: Column_GID
        required: true
        type: string
      responses:
        "200":
          description: ""
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - column
    get:
      description: Retrieves column by parent board and column UUIDs
      operationId: column-retrieve-one
      parameters:
      - description: UUID of parent board
        in: path
        name: Board_GID
        required: true
        type: string
      - description: UUID of column
        in: path
        name: Column_GID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/single-column-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - column
    put:
      description: Updates column by parent board and column UUIDs
      operationId: column-update
      parameters:
      - description: UUID of parent board
        in: path
        name: Board_GID
        required: true
        type: string
      - description: UUID of column
        in: path
        name: Column_GID
        required: true
        type: string
      - description: Title of column
        in: body
        name: Title
        schema:
          type: string
      responses:
        "200":
          $ref: '#/responses/single-column-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - column
...
  /api/register:
    post:
      description: Registers a new member using the supplied username, email, and
        password
      operationId: register
      parameters:
      - example: andrew
        in: formData
        name: Username
        required: true
        type: string
      - example: user@email.com
        in: formData
        name: Email
        required: true
        type: string
      - example: password1234
        in: formData
        name: Password
        required: true
        type: string
      responses:
        "200":
          description: ""
        "400":
          $ref: '#/responses/error-response'
      tags:
      - authentication
...
produces:
- application/json
responses:
  error-response:
    description: ""
    schema:
      $ref: '#/definitions/Error'
...
  multi-column-response:
    description: ""
    schema:
      items:
        $ref: '#/definitions/TaskColumn'
      type: array
...
  single-column-response:
    description: ""
    schema:
      $ref: '#/definitions/TaskColumn'
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"